+++ src/Core/Dfe.Complete.Infrastructure/CommandServices/UserWriteRepository.cs
@@ -1,0 +1,16 @@
+using Dfe.Complete.Application.Users.Interfaces;
+using Dfe.Complete.Domain.Entities;
+using Dfe.Complete.Infrastructure.Database;
+
+namespace Dfe.Complete.Infrastructure.CommandServices;
+
+internal class UserWriteRepository(CompleteContext context) : IUserWriteRepository
+{
+    private readonly CompleteContext _context = context;
+
+    public async Task CreateUserAsync(User user, CancellationToken cancellationToken)
+    {
+        await _context.Users.AddAsync(user, cancellationToken);
+        await _context.SaveChangesAsync(cancellationToken);
+    }
+}